NSA-1699 refactoring mapping
